+++ meetupguider-api/src/MeetupGuilder.Entities/Migrations/MeetupGuiderDbContextModelSnapshot.cs
@@ -80,8 +80,8 @@ namespace MeetupGuilder.Entities.Migrations
                     b.Property<string>("EventId")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("GroupId")
-                        .HasColumnType("int");
+                    b.Property<long>("GroupId")
+                        .HasColumnType("bigint");
 
                     b.Property<float>("Lat")
                         .HasColumnType("real");
